create transcribe.js route
