+++ lib/files.js
