+++ lib/stats.js
